+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## [Unreleased]
 
+* Fixed an issue that would stop the `rename` query parameter from being handled correctly.
 * Fixed issue that could cause package download URL's to not resolve properly.
 * Improved search handling to support spaces among other stop words to still resolve the expected search query.
 * Deprecate `/api/packages/:packageName/versions/:versionName/events/uninstall`. This endpoint no longer has any effect, but will still return a successful query to avoid user impact.